Edit - Creating index.js and adding all my methods (displayWorks, setActiveButton etc..) & v1 of separating my filter.js file
