+++ FrontEnd/index.html
@@ -9,8 +9,8 @@
     <link href="https://fonts.googleapis.com/css2?family=Syne:wght@700;800&family=Work+Sans&display=swap" rel="stylesheet">
     <meta name="description" content="">
     <link rel="stylesheet" href="./assets/style.css">
-	<script type="module" src="./js/displayWorks.js"></script> 
-	<script type="module" src="./js/filter.js"></script> 
+	<script type="module" src="./js/filteredWorksByCategory.js"></script> 
+    <script type="module" src="./index.js"></script>
 </head>
 <body>
 <header>
