+++ FrontEnd/index.js
@@ -1,0 +1,36 @@
+// Afficher les catégories (1)
+
+
+// Afficher les works (2)
+export function displayWorks(works) {
+    const gallery = document.querySelector('.gallery');
+    gallery.innerHTML = '';
+
+    works.forEach(work => {
+        const figure = document.createElement('figure');
+        const img = document.createElement('img');
+        const figcaption = document.createElement('figcaption');
+
+        img.src = work.imageUrl;
+        img.alt = work.title;
+        figcaption.textContent = work.title;
+        
+        figure.appendChild(img);
+        figure.appendChild(figcaption);
+        gallery.appendChild(figure);
+    });
+}
+
+// Gérer le filtre (3)
+
+// Afficher l'état du button (4) 
+export function setActiveButton(activeButton) {
+
+    const filter = document.querySelector('.filter');
+
+    const buttons = filter.querySelectorAll('button');
+    buttons.forEach(button => {
+        button.classList.remove('active');
+    });
+    activeButton.classList.add('active');
+}
