--- FrontEnd/js/displayWorks.js
@@ -1,18 +1,0 @@
-export function displayWorks(works) {
-    const gallery = document.querySelector('.gallery');
-    gallery.innerHTML = '';
-
-    works.forEach(work => {
-        const figure = document.createElement('figure');
-        const img = document.createElement('img');
-        const figcaption = document.createElement('figcaption');
-
-        img.src = work.imageUrl;
-        img.alt = work.title;
-        figcaption.textContent = work.title;
-        
-        figure.appendChild(img);
-        figure.appendChild(figcaption);
-        gallery.appendChild(figure);
-    });
-}
